Migrate to vitest
