--- packages/esm-patient-procedures-app/jest.config.js
@@ -1,3 +1,0 @@
-const rootConfig = require('../../jest.config.js');
-
-module.exports = rootConfig;
