+++ packages/esm-patient-procedures-app/src/components/concept-search-field/concept-search-field.component.tsx
@@ -1,0 +1,93 @@
+import { InlineLoading, Layer, Search, Tile } from '@carbon/react';
+import React from 'react';
+import { ResponsiveWrapper } from '@openmrs/esm-framework';
+import { type ConceptReference } from '../../types';
+import { type useConceptSearchField } from '../../procedures.resource';
+import styles from './concept-search-field.scss';
+import { useTranslation } from 'react-i18next';
+
+export const ConceptSearchField = ({
+  label,
+  placeholder,
+  field,
+  onChange,
+}: {
+  label: string;
+  placeholder: string;
+  field: ReturnType<typeof useConceptSearchField>;
+  onChange: (selectedConcept: ConceptReference) => void;
+}) => {
+  return (
+    <>
+      <ResponsiveWrapper>
+        <Search
+          labelText={label}
+          placeholder={placeholder}
+          onChange={(e) => field.setSearchTerm(e.target.value)}
+          onClear={field.clear}
+          value={field.selectedConcept ? field.selectedConcept.display : field.searchTerm}
+        />
+      </ResponsiveWrapper>
+
+      <ConceptSearchResults
+        isSearching={field.isSearching}
+        searchResults={field.searchResults}
+        selectedItem={field.selectedConcept}
+        value={field.searchTerm}
+        onSelect={(result) => {
+          field.setSelectedConcept(result);
+          field.setSearchTerm('');
+          onChange(result);
+        }}
+      />
+    </>
+  );
+};
+
+interface ConceptSearchResultsProps {
+  isSearching: boolean;
+  onSelect: (result: ConceptReference) => void;
+  searchResults: Array<ConceptReference>;
+  selectedItem: ConceptReference;
+  value: string;
+}
+
+const ConceptSearchResults = ({
+  isSearching,
+  onSelect,
+  searchResults,
+  selectedItem,
+  value,
+}: ConceptSearchResultsProps) => {
+  const { t } = useTranslation();
+
+  if (!value || selectedItem) {
+    return null;
+  }
+
+  if (isSearching) {
+    return <InlineLoading className={styles.loader} description={t('searching', 'Searching') + '...'} />;
+  }
+
+  if (searchResults?.length > 0) {
+    return (
+      <ul className={styles.resultsList}>
+        {searchResults.map((result) => (
+          <li className={styles.resultItem} key={result.uuid} onClick={() => onSelect(result)} role="menuitem">
+            {result.display}
+          </li>
+        ))}
+      </ul>
+    );
+  }
+
+  return (
+    <Layer>
+      <Tile className={styles.emptyResults}>
+        <span>
+          {t('noResultsFor', 'No results for')} <strong>"{value}"</strong>
+        </span>
+      </Tile>
+    </Layer>
+  );
+};
