+++ packages/esm-patient-procedures-app/src/components/concept-search-field/concept-search-field.scss
@@ -1,0 +1,39 @@
+@use '@carbon/colors';
+@use '@carbon/layout';
+@use '@carbon/type';
+@use '@openmrs/esm-styleguide/src/vars' as *;
+
+.resultItem {
+  padding: layout.$spacing-04;
+  border-top: 0.0625rem solid $grey-2;
+
+  &:first-of-type {
+    border-top: none;
+  }
+
+  &:last-of-type {
+    border-bottom: none;
+  }
+}
+
+.resultsList {
+  background-color: $ui-02;
+  max-height: 14rem;
+  overflow-y: auto;
+  border: 1px solid $ui-03;
+  list-style: none;
+  padding: 0;
+  margin: 0;
+
+  li:hover {
+    background-color: $ui-03;
+    cursor: pointer;
+  }
+}
+
+.emptyResults {
+  @include type.type-style('body-compact-01');
+  color: $text-02;
+  min-height: layout.$spacing-05;
+  border: 1px solid $ui-03;
+}
