+++ packages/esm-patient-procedures-app/src/components/date-time-field/date-time-field.component.tsx
@@ -4,7 +4,7 @@ import { SelectItem, TimePicker, TimePickerSelect } from '@carbon/react';
 import { Controller, type Control } from 'react-hook-form';
 import { OpenmrsDatePicker, ResponsiveWrapper } from '@openmrs/esm-framework';
 import { type ProceduresFormSchema } from '../../workspaces/procedures-form/procedures-form.workspace';
-import styles from './procedures-form.scss';
+import styles from './date-time-field.scss';
 
 const TIME_PATTERN = /^(1[0-2]|0?[1-9]):([0-5]\d)$/;
