+++ packages/esm-patient-procedures-app/src/components/date-time-field/date-time-field.scss
@@ -1,0 +1,8 @@
+@use '@carbon/layout';
+
+.dateTimeFieldGroup {
+  display: grid;
+  grid-template-columns: 2fr 1fr;
+  gap: layout.$spacing-05;
+  align-items: end;
+}
