+++ packages/esm-patient-procedures-app/src/components/detailed-summary/procedures-detailed-summary.component.test.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { vi, describe, it, expect } from 'vitest';
 import { screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import {
@@ -13,13 +14,13 @@ import { mockPatient, renderWithSwr, waitForLoadingToFinish } from 'tools';
 import ProceduresDetailedSummary from './procedures-detailed-summary.component';
 import { mockProceduresResponse } from '__mocks__';
 
-jest.mock('./procedures-action-menu.component', () => ({
-  ProceduresActionMenu: jest.fn().mockReturnValue(null),
+vi.mock('./procedures-action-menu.component', () => ({
+  ProceduresActionMenu: vi.fn().mockReturnValue(null),
 }));
 
-const mockUseConfig = jest.mocked(useConfig<ConfigObject>);
-const mockOpenmrsFetch = jest.mocked(openmrsFetch);
-const mockLaunchWorkspace2 = jest.mocked(launchWorkspace2);
+const mockUseConfig = vi.mocked(useConfig<ConfigObject>);
+const mockOpenmrsFetch = vi.mocked(openmrsFetch);
+const mockLaunchWorkspace2 = vi.mocked(launchWorkspace2);
 
 mockUseConfig.mockReturnValue({
   ...getDefaultsFromConfigSchema(configSchema),
