+++ packages/esm-patient-procedures-app/src/index.ts
@@ -2,10 +2,10 @@ import { getAsyncLifecycle, getSyncLifecycle, defineConfigSchema } from '@openmr
 import { createDashboardLink } from '@openmrs/esm-patient-common-lib';
 import { configSchema } from './config-schema';
 import { dashboardMeta } from './dashboard.meta';
-import proceduresOverviewComponent from './components/procedures-overview.component';
-import proceduresDetailedSummaryComponent from './components/procedures-detailed-summary.component';
-import proceduresFormWorkspaceComponent from './components/procedures-form.workspace';
 import { moduleName } from './constants';
+import ProceduresOverview from './components/overview/procedures-overview.component';
+import ProceduresDetailedSummary from './components/detailed-summary/procedures-detailed-summary.component';
+import ProceduresFormWorkspace from './workspaces/procedures-form/procedures.form';
 
 const options = {
   featureName: 'patient-procedures-app',
@@ -19,17 +19,17 @@ export function startupApp() {
 }
 
 // Extensions
-export const proceduresOverview = getSyncLifecycle(proceduresOverviewComponent, options);
+export const proceduresOverview = getSyncLifecycle(ProceduresOverview, options);
 
-export const proceduresDetailedSummary = getSyncLifecycle(proceduresDetailedSummaryComponent, options);
+export const proceduresDetailedSummary = getSyncLifecycle(ProceduresDetailedSummary, options);
 
 export const proceduresDashboardLink =
   // t('Procedures', 'Procedures')
   getSyncLifecycle(createDashboardLink({ ...dashboardMeta }), options);
 
-export const proceduresFormWorkspace = getSyncLifecycle(proceduresFormWorkspaceComponent, options);
+export const proceduresFormWorkspace = getSyncLifecycle(ProceduresFormWorkspace, options);
 
 export const procedureDeleteConfirmationDialog = getAsyncLifecycle(
-  () => import('./components/delete-procedure.modal'),
+  () => import('./modals/delete-procedure/delete-procedure.modal'),
   options,
 );
